requirements.txt
pandas
openpyxl
praw
python-dotenv
requests
